Fixed an anomaly between GCC and Visual Studio Compiler
It would seem that the visual Studio compiler does not enjoy being passed a primitive data type and then have a pointer point at that data type. It returns incoherent data. As such the passed int must be used to create an new int(int value) which creates a pointer and a new instance in memory. 

This behavior was not observed when using GCC on Linux.